GUEST functions for Admin Dashboard finished
GUEST role functionality works now, GUEST can add, edit and delete their
own products. Alerts need extension to make them movable and/or
closable, its in the way of products link when you enter admin dashboard
as a GUEST.

! All Cypress tests broken apart from auth, fix later!

+++ src/features/admin/admin-dashboard/components/categories/categories-dashboard.component.tsx
@@ -5,6 +5,7 @@ import { Paginator } from '../../../../forms/paginator';
 import { selectCategories } from '../../../../category/categorySlice';
 import { fetchCategories } from '../../../../category/api';
 import { remove as removeCategory } from '../../../../category/thunks';
+import { selectRole } from '../../../../user/selectors';
 
 type SelectionType = {
   id: string;
@@ -13,6 +14,7 @@ type SelectionType = {
 function CategoriesDashboard(): JSX.Element {
   const dispatch = useDispatch();
   const categories = useSelector(selectCategories);
+  const role = useSelector(selectRole);
   const [currentPage, setCurrentPage] = React.useState<number>(0);
   const [view, setView] = React.useState<number>(10);
   const [input, setInput] = React.useState({ search: '' });
@@ -28,6 +30,10 @@ function CategoriesDashboard(): JSX.Element {
   }
 
   function deleteHandler(e: React.MouseEvent<HTMLButtonElement>) {
+    if (role === 'GUEST') {
+      alert(`You dont have sufficient rights to do a submission`);
+      return;
+    }
     selections.length <= 1
       ? confirm('Are you sure you want to delete this product?') &&
         dispatch(removeCategory(JSON.parse(e.currentTarget.value)))
